Fixed an issue with raphael
Raphael uses document.location to generate the fill url, which
points to an existing filter id in the page using a hash (#). Since
Angluarjs is using a hash in the URL, this conflicts with Raphael
generator and canvas renders black. Enabling html5 mode fixed the
problem.

Also removed the unused anglepicker directive.